requirements.txt
fastapi
uvicorn
sqlalchemy
pyodbc
psycopg2-binary
python-jose
passlib[bcrypt]
python-multipart
email-validator
openai
python-dotenv
requests
pydantic